Separated debug hooks
